requirements.txt
flask
python-dotenv
flask_sqlalchemy
flask-security
faker
pandas
email_validator